Defer service provider
Since you are only wanting this provider to do stuff when you are on the console, it can be deferred. All providers are loaded in the console environment regardless of whether they are deferred or not. This will prevent CollisionServiceProvider from ever being loaded at all in the HTTP environment.